requirements.txt
click==8.1.8
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
typer==0.15.2
typing_extensions==4.13.1
